Move Typed tests model to separate namespace
